backend/requirements.txt
fastapi>=0.116,<1.0
PyJWT[crypto]>=2.10,<3.0
python-dotenv>=1.1,<2.0
requests>=2.32,<3.0
uvicorn[standard]>=0.35,<1.0
